refactor: styles user move to separate file
